Move generic functions to module_utils #125
